Refactor : 게스트 입장 기능 custom hooks로 리팩토링
